legacy/requirements.txt
requests~=2.28.1
requests-html
flask~=2.1.3
setuptools~=57.0.0
pysocks
msgpack
